Separate filter functions into independent functions
